Grokie/requirements.txt
livekit-agents[xai]
livekit
python-dotenv

